+++ packages/arbor-store/src/index.ts
@@ -8,6 +8,6 @@ export { NodeHandler } from "./handlers/NodeHandler"
 export * from "./MutationEngine"
 export { Path } from "./Path"
 export * from "./ScopedStore"
-export { Subscribers } from "./Subscribers"
+export { Subscribers } from "./Subscriptions"
 export type * from "./types"
 export * from "./utilities"
