+++ packages/arbor-store/tests/Subscribers.test.ts
@@ -2,13 +2,13 @@ import { describe, expect, it, vi } from "vitest"
 
 import { Arbor } from "../src/Arbor"
 import { Path } from "../src/Path"
-import { Subscribers } from "../src/Subscribers"
+import { Subscriptions } from "../src/Subscriptions"
 
-describe("Subscribers", () => {
+describe("Subscriptions", () => {
   it("notifies subscribers of mutation events", () => {
     const subscriber1 = vi.fn()
     const subscriber2 = vi.fn()
-    const subscribers = new Subscribers()
+    const subscribers = new Subscriptions()
     const mutationEvent = {
       store: new Arbor({}),
       mutationPath: Path.root,
@@ -28,7 +28,7 @@ describe("Subscribers", () => {
   it("does not notify subscribers that have canceled their subscription", () => {
     const subscriber1 = vi.fn()
     const subscriber2 = vi.fn()
-    const subscribers = new Subscribers()
+    const subscribers = new Subscriptions()
     const mutationEvent = {
       store: new Arbor({}),
       mutationPath: Path.root,
@@ -49,7 +49,7 @@ describe("Subscribers", () => {
 
   describe("#size", () => {
     it("returns the number of subscribers registered thus far", () => {
-      const subscribers = new Subscribers()
+      const subscribers = new Subscriptions()
 
       expect(subscribers.size).toBe(0)
       subscribers.subscribe(vi.fn())
